# Conflicts:
#	.openzeppelin/unknown-97.json
#	deploy/interfaces/config.ts
#	hardhat.config.ts
#	package.json
#	test/helpers/assert.ts
#	test/helpers/deploy.ts
#	yarn.lock

+++ contracts/6/protocol/apis/pancake/PancakeLibrary.sol
@@ -2,11 +2,11 @@ pragma solidity >=0.5.0;
 
 import "@openzeppelin/contracts-ethereum-package/contracts/math/SafeMath.sol";
 
-import "@pancakeswap-libs/pancake-swap-core/contracts/interfaces/IPancakePair.sol";
-import "@pancakeswap-libs/pancake-swap-core/contracts/interfaces/IPancakeFactory.sol";
+import "../../interfaces/IPancakePair.sol";
+import "../../interfaces/IPancakeFactory.sol";
 
 library PancakeLibrary {
-  using SafeMath for uint;
+  using SafeMath for uint256;
 
   // returns sorted token addresses, used to handle return values from pairs sorted in this order
   function sortTokens(address tokenA, address tokenB) internal pure returns (address token0, address token1) {
@@ -16,62 +16,90 @@ library PancakeLibrary {
   }
 
   // calculates the CREATE2 address for a pair without making any external calls
-  function pairFor(address factory, address tokenA, address tokenB) internal view returns (address pair) {
+  function pairFor(
+    address factory,
+    address tokenA,
+    address tokenB
+  ) internal view returns (address pair) {
     return IPancakeFactory(factory).getPair(tokenA, tokenB); // For easy testing
   }
 
   // fetches and sorts the reserves for a pair
-  function getReserves(address factory, address tokenA, address tokenB) internal view returns (uint reserveA, uint reserveB) {
-    (address token0,) = sortTokens(tokenA, tokenB);
+  function getReserves(
+    address factory,
+    address tokenA,
+    address tokenB
+  ) internal view returns (uint256 reserveA, uint256 reserveB) {
+    (address token0, ) = sortTokens(tokenA, tokenB);
     pairFor(factory, tokenA, tokenB);
-    (uint reserve0, uint reserve1,) = IPancakePair(pairFor(factory, tokenA, tokenB)).getReserves();
+    (uint256 reserve0, uint256 reserve1, ) = IPancakePair(pairFor(factory, tokenA, tokenB)).getReserves();
     (reserveA, reserveB) = tokenA == token0 ? (reserve0, reserve1) : (reserve1, reserve0);
   }
 
   // given some amount of an asset and pair reserves, returns an equivalent amount of the other asset
-  function quote(uint amountA, uint reserveA, uint reserveB) internal pure returns (uint amountB) {
+  function quote(
+    uint256 amountA,
+    uint256 reserveA,
+    uint256 reserveB
+  ) internal pure returns (uint256 amountB) {
     require(amountA > 0, "PancakeLibrary: INSUFFICIENT_AMOUNT");
     require(reserveA > 0 && reserveB > 0, "PancakeLibrary: INSUFFICIENT_LIQUIDITY");
     amountB = amountA.mul(reserveB) / reserveA;
   }
 
   // given an input amount of an asset and pair reserves, returns the maximum output amount of the other asset
-  function getAmountOut(uint amountIn, uint reserveIn, uint reserveOut) internal pure returns (uint amountOut) {
+  function getAmountOut(
+    uint256 amountIn,
+    uint256 reserveIn,
+    uint256 reserveOut
+  ) internal pure returns (uint256 amountOut) {
     require(amountIn > 0, "PancakeLibrary: INSUFFICIENT_INPUT_AMOUNT");
     require(reserveIn > 0 && reserveOut > 0, "PancakeLibrary: INSUFFICIENT_LIQUIDITY");
-    uint amountInWithFee = amountIn.mul(998);
-    uint numerator = amountInWithFee.mul(reserveOut);
-    uint denominator = reserveIn.mul(1000).add(amountInWithFee);
+    uint256 amountInWithFee = amountIn.mul(998);
+    uint256 numerator = amountInWithFee.mul(reserveOut);
+    uint256 denominator = reserveIn.mul(1000).add(amountInWithFee);
     amountOut = numerator / denominator;
   }
 
   // given an output amount of an asset and pair reserves, returns a required input amount of the other asset
-  function getAmountIn(uint amountOut, uint reserveIn, uint reserveOut) internal pure returns (uint amountIn) {
+  function getAmountIn(
+    uint256 amountOut,
+    uint256 reserveIn,
+    uint256 reserveOut
+  ) internal pure returns (uint256 amountIn) {
     require(amountOut > 0, "PancakeLibrary: INSUFFICIENT_OUTPUT_AMOUNT");
     require(reserveIn > 0 && reserveOut > 0, "PancakeLibrary: INSUFFICIENT_LIQUIDITY");
-    uint numerator = reserveIn.mul(amountOut).mul(1000);
-    uint denominator = reserveOut.sub(amountOut).mul(998);
+    uint256 numerator = reserveIn.mul(amountOut).mul(1000);
+    uint256 denominator = reserveOut.sub(amountOut).mul(998);
     amountIn = (numerator / denominator).add(1);
   }
 
   // performs chained getAmountOut calculations on any number of pairs
-  function getAmountsOut(address factory, uint amountIn, address[] memory path) internal view returns (uint[] memory amounts) {
+  function getAmountsOut(
+    address factory,
+    uint256 amountIn,
+    address[] memory path
+  ) internal view returns (uint256[] memory amounts) {
     require(path.length >= 2, "PancakeLibrary: INVALID_PATH");
-    amounts = new uint[](path.length);
+    amounts = new uint256[](path.length);
     amounts[0] = amountIn;
-    for (uint i; i < path.length - 1; i++) {
-      (uint reserveIn, uint reserveOut) = getReserves(factory, path[i], path[i + 1]);
+    for (uint256 i; i < path.length - 1; i++) {
+      (uint256 reserveIn, uint256 reserveOut) = getReserves(factory, path[i], path[i + 1]);
       amounts[i + 1] = getAmountOut(amounts[i], reserveIn, reserveOut);
     }
   }
 
   // performs chained getAmountIn calculations on any number of pairs
-  function getAmountsIn(address factory, uint amountOut, address[] memory path) internal view returns (uint[] memory amounts) {
+  function getAmountsIn(
+    address factory,
+    uint256 amountOut,
+    address[] memory path
+  ) internal view returns (uint256[] memory amounts) {
     require(path.length >= 2, "PancakeLibrary: INVALID_PATH");
-    amounts = new uint[](path.length);
+    amounts = new uint256[](path.length);
     amounts[amounts.length - 1] = amountOut;
-    for (uint i = path.length - 1; i > 0; i--) {
-      (uint reserveIn, uint reserveOut) = getReserves(factory, path[i - 1], path[i]);
+    for (uint256 i = path.length - 1; i > 0; i--) {
+      (uint256 reserveIn, uint256 reserveOut) = getReserves(factory, path[i - 1], path[i]);
       amounts[i - 1] = getAmountIn(amounts[i], reserveIn, reserveOut);
     }
   }
